feat(game) Implement card component
Implement logic for card component.

The card component takes as input a card ID, an 'active' flag (true by default),
and a state (FaceDown by default, alternativley FaceUp or Hidden). It emits its card ID when clicked.

The game component should be possible for determining the state and activity
of its cards (children), as knowledge of other cards' state is required to
determine how to change said card's state in response to a click (if at all).